Update docker setup
The start.sh script now builds only the adapter outside of the image
building process. This reduces the amount of work that is needed to run
tests with different builds.

Added stop.sh (shuts off the test environment) and build.sh (builds the
adapter) scripts.

Fixed a build failure on CentOS 7 caused by the initialization of the
struct sigaction. A workaround is to use memset to clear it.