MCOL-1594: Add fast shutdown
The adapter will now stop as soon as possible. This is done by using a
dedicated shutdown thread that simply waits for termination signals and
starts the shutdown when one is received.

Also made the signal handler async-signal safe and altered the immediate
shutdown code to wait for at least two signals before stopping the
process.