Fix shutdown on error
When the adapter was stopped due to an error, the shutdown thread would
not exit due to the flag not being set.

Also added a mutex to guard the use and modification of the context list
and changed from vector to list to make sure iterators aren't affected by
removals from the list.