MCOL-1661: Add data transformation
The data can now be transformed into DML statements that keep the
ColumnStore instance up to date with the MariaDB master. This is not a
performant solution as ColumnStore is not optimal for OLTP type loads but
it is an adequate starting point. Once MCOL-1662 is solved, INSERT
performance will improve greatly but other DMLs will still remain
inefficient.